refactor: move GetImageConfigFromImage to new pkg
